Preserve native SQLite data keys across history changes

+++ crates/uqa-storage-sqlite/src/mvcc/native/tests/graph_lookup/path_migration.rs
@@ -149,7 +149,7 @@ fn native_path_lookup_upgrade_preserves_sources_and_old_selectors_in_every_file_
             }
             assert_eq!(
                 dump(&connection, "SELECT format FROM _uqa_mvcc_native_format"),
-                vec![vec![rusqlite::types::Value::Integer(8)]]
+                vec![vec![rusqlite::types::Value::Integer(9)]]
             );
             assert_eq!(
                 SQLiteRecordStore::for_native(&connection, &control)
